Spring now talking to postgres sql

+++ src/main/java/com/redhat/analytics/jiminy/htmlserver/service/RatingService.java
@@ -4,45 +4,19 @@ import com.redhat.analytics.jiminy.htmlserver.model.RatingsDAO;
 import java.util.List;
 
 public interface RatingService {
-//    interface ValidationRules {
-//        int max();
-//
-//        int min();
-//    }
-//
-//    // Validation Rules. Plan on moving this somewhere else but for here its
-//    // here.
-//    enum ValidUser implements ValidationRules {
-//        USERNAME {
-//            public int max() {
-//                return 20;
-//            }
-//
-//            public int min() {
-//                return 5;
-//            }
-//        },
-//        PASSWORD {
-//            public int max() {
-//                return 10;
-//            }
-//
-//            public int min() {
-//                return 5;
-//            }
-//        };
-//    }
-
+ 
     RatingsDAO create(RatingsDAO user);
 
-    RatingsDAO delete(Integer id);
+    RatingsDAO delete(long id);
 
     List<RatingsDAO> findAll();
 
-    RatingsDAO findById(Integer id);
+    RatingsDAO findById(long id);
 
     RatingsDAO update(RatingsDAO users);
 
     void deleteAll();
+
+  
 
 }
